refactor: home ExportState with the macro export SPI
ExportState is public runtime SPI called only by @Syncable-generated export()
code (the macro emits ExportState.enter/leave) — the same category as
exportEncodeValue/exportSetValue, which live in SyncableMacroSupport.swift.
Fold it in beside them rather than standing alone.

(ParentScopedModel stays its own file: it is not macro-generated — it is a
public protocol refining SyncUpdatableModel that consumers conform to by hand,
so it belongs with the model-protocol family, not the macro support.)

Pure move: SwiftSync 172 (+48), DemoCore 43/43 green.
